Finish porting permission and license pages
